Add rating

+++ src/admin/admin.service.ts
@@ -1,13 +1,16 @@
-import { Injectable } from '@nestjs/common'
+import { Injectable, NotFoundException } from '@nestjs/common'
 import { InjectRepository } from '@nestjs/typeorm'
 import { FileEntity } from '../files/entities/file.entity'
 import { Repository } from 'typeorm'
+import { CommentEntity } from '../comments/entities/comment.entity'
 
 @Injectable()
 export class AdminService {
   constructor(
     @InjectRepository(FileEntity)
     private fileEntityRepository: Repository<FileEntity>,
+    @InjectRepository(CommentEntity)
+    private commentEntityRepository: Repository<CommentEntity>,
   ) {}
 
   getFilesPending() {
@@ -55,13 +58,27 @@ export class AdminService {
     await qb.update().set({ restricted: 'private', reject: true }).execute()
   }
 
-  async delete(id: string) {
+  async deleteFile(id: string) {
     const qb = this.fileEntityRepository.createQueryBuilder('file')
 
     qb.where('id = :id AND restricted = :restricted', {
       id: id,
       restricted: 'public',
     })
+
+    await qb.delete().execute()
+  }
+
+  async deleteComment(id: string) {
+    const qb = this.commentEntityRepository.createQueryBuilder('comment')
+
+    const comment = qb.where('id = :id', {
+      id: id,
+    })
+
+    if (comment === undefined) {
+      throw new NotFoundException('Comment not found')
+    }
 
     await qb.delete().execute()
   }
